refactor: move app/ into src/app for a single source root
Next App Router supports src/app; consolidating under baseUrl=src removes the
second root. Rewrote cross-root relative imports to baseUrl-bare/sibling specifiers;
updated tailwind content, tsconfig include, and lint globs. Build resolves all routes
from src/app; typecheck/lint/format/build pass.
