+++ package.json
@@ -15,8 +15,8 @@
     "build": "next build",
     "preview": "pnpm dlx serve out -l 4173",
     "typecheck": "tsc --noEmit",
-    "lint": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"app/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\"",
-    "lint:fix": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"app/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\" --fix",
+    "lint": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\"",
+    "lint:fix": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\" --fix",
     "test": "vitest run --reporter=default",
     "test:coverage": "vitest run --coverage.enabled=true --coverage.provider=istanbul --coverage.reporter=lcov --coverage.reporter=html --coverage.reporter=text",
     "format": "prettier -w .",
@@ -47,8 +47,8 @@
     "node:dev": "next dev -p 1420",
     "node:build": "next build",
     "node:typecheck": "tsc --noEmit",
-    "node:lint": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"app/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\"",
-    "node:lint:fix": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"app/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\" --fix",
+    "node:lint": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\"",
+    "node:lint:fix": "eslint --max-warnings=10 \"src/**/*.{ts,tsx,js}\" \"tests/**/*.{ts,tsx,js}\" --fix",
     "node:test": "vitest run --reporter=default",
     "node:test:coverage": "vitest run --coverage.enabled=true --coverage.provider=istanbul --coverage.reporter=lcov --coverage.reporter=html --coverage.reporter=text",
     "node:format": "prettier -w .",
